Split JavaScriptWebWorker and JavaScriptImportScripts tests, increase coverage.
