+++ testdata/relations/JavaScriptImportScripts/seq/bar.js
