+++ testdata/relations/JavaScriptImportScripts/seq/foo.js
