+++ testdata/relations/JavaScriptImportScripts/seq/index.html
