+++ testdata/relations/JavaScriptImportScripts/seq/worker.js
