+++ testdata/relations/JavaScriptImportScripts/simple/bar.js
@@ -1,0 +1,1 @@
+console.log('bar');
