+++ testdata/relations/JavaScriptImportScripts/simple/foo.js
@@ -1,0 +1,1 @@
+console.log('foo');
