+++ testdata/relations/JavaScriptImportScripts/simple/index.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+    <head></head>
+    <body>
+        <script>
+            new Worker('worker.js');
+        </script>
+    </body>
+</html>
