+++ testdata/relations/JavaScriptImportScripts/simple/quux.js
@@ -1,0 +1,1 @@
+console.log('quux');
