+++ testdata/relations/JavaScriptImportScripts/simple/worker.js
@@ -1,0 +1,3 @@
+importScripts('foo.js', 'bar.js');
+
+self.importScripts('quux.js');
